Fix bad data aggregation algorithm
The data aggregation algorithm did not do what it is expected to do.
Indeed, it waits for each coroutine to terminate in order to retrieve data.
The application must not do that.

This commit fixes that, so the application can start retrieving data
if and only if there is at least one coroutine that is sending data.
It does not matter if this coroutine finished its job or not.