Improve the aggregation of results
The previous aggregation algorithm has a big problem.
It actively waits for the services to finish its job, so it can retrieve data
(each service is handled by a dedicated coroutine).

This active waiting has consequences on the CPU usage
(on my computer, 50% of my CPU [2 cores] are used just for that).

This commit fixes that by changing the model of aggregation.

> Old model

In the old model, the aggregator launches services in parallel (1 service = 1 coroutine)
those services have the role of producers, so each associated coroutine provides a channel to dans data in.
The agregator has the role of the consumer and has access to those channels.

The consumer checks for each channel if it contains data in order to start retrieving data.
This consumer will stop the aggregation when every services finished their job,
i.e. if every channels are closed and has no more data to retrieve.

In the first implementation, the consumer starts retrieving data when each service is terminated,
and not when the data it sends is available.
This unexpected behaviour is sequential!

The second implementation is not sequential and start getting data when they are
available, even if those services are not terminated. However, the aggregator actively waits.

> New model

This model is still using the producer-consumer pattern, but instead of having
one coroutine that send data to a specific channel. It send data to a channel
provided by the aggregator. This channel is shared between those producers.
Those coroutine

The main difference/advantage of this model is that the aggregator can **passively** wait,
so it can get data when they are available.
However, there is a potential retention on the producer side.
It's acceptable!