Fix bad usage of class mocking in tests
Using @BeforeTest for mocking is bad, because your test will
mock the same function call at each test. The mock is already set
at the creation of the test object, so having multiple function mockings
are redundant and a waste of time.

This commit fixes that by setting everything once in the class mock creation.