Create the rest controller once and use it
> Before this commit

The RestController was created for each request in order to handle the request.
After that, the object is not used and has to be collected by the garbage collector (GC).

> What this commit change

The rest controller is created once when the application module is created.

> Why?

Creating a object with a lifecycle limited to a request handling is very
memory consuming. But instantiating the controller once and use it multiple times
consume less memory.