Implement a rate limiter for Arxiv
The use of the legacy arXiv API implies that Gakusci must use the API
without voilating or circumventing rate limits.

In order to avoid getting banned by the service, the rate limit is implemented
in the arXiv client, so the application can use the service without problem.

> Why not using the arXiv API Gateway?

Even though the arXiv API Gateway has higher rate limits, an authentication is needed.
Gakusci is built in order to provide more visibility to public and open APIs that can be
used without getting authenticated or providing any API key/token.
That is the reason why Gakusci uses the legacy API instead of that.

More information: https://arxiv.org/help/api/tou#rate-limits