Refactorize the search aggregation
> Before

The application sent a list of entries.

Why is it a problem ?

Sending a list of results is not the problem. The problem is that
it did not send information about the total results.

> Now

Tha pplication send a list of rentries, the number of results (it may be different from the number of entries)
and the start index (very useful for pagination).