Read the feed properly by using a custom feed reader
> Problem

I need to read fields from the opensearch namespace, and
get the existing fields properly without wasting my time.

Rome can retrieves any fields of the feed properly.
It can also retrieve opensearch values thanks to a dedicated module.
However, this module is not integrated in the last version of the library.

I tried to implement a custom module thet does the job, but I had several problems.
Besides, the documentation is not really up-to-date.

I also noticed that even if Rome is popular, there are few contributions,
and the last version was released in 2018.

So I decided to check if a proper and modern syndication feed reader can do
the job without pain.

> Solution

I find a tool that can do what I need to do: Syndication (Kotlin)

Syndication (Kotlin) allows you to get Atom/RSS feed properly
from an external source (Arxiv). This tool uses Okhttp to interact with ATOM/RSS server.
Syndication is built so you don't have to take care about the reading implementation.
You just need to specify an interface that defines the reader and annotate it,
so you can choose to get an implementation of:
    - an RSS feed reader.
    - an ATOM feed reader.

You can also define your own feed reader. This is very useful when you want to
retrieve specific element from your feed for specific usages.

The philosophy behind Syndication is based on some principles from SimpleXML, OkHttp.
It also can work within coroutines, which is very important here, because the webapp
is based one coroutines.

Links:

- https://github.com/ouattararomuald/syndication
- https://github.com/square/okhttp
- https://github.com/codemonstur/simplexml