Display results per page of `n` entries
It is possible to display results per page.
The number of elements to display per page can be specified in the REST API.

> Why?

Gakusci can display at most 1000 results per request in order to
prevent performance issues, even if there is more results.

However, some clients need to display results per page.
That is why this commit adds the implementation of pagination.