Implement the Open Library client [Books]
The documentation of the Open Library API is not complete
because some properties are not properly documented.

The JSON deserialization could not work properly as it does
with HAL. Indeed, the web-service does not provide information about
the type of content returned. So I had to retrieve the string and "manually"
deserialize it to the associated object (OpenLibraryResponse).