Adapt the maximum number of results depending on the page to display [web]
The 'max_results' query value (the 'rows' property in the application logic)
is not very useful in the web context of the application. It is mainly used to
retrieved a maximal number of results per external services.

That is why its value is adapted depending on the 'start' query parameter, so
the web application can retrieve more results on demand, especially if those results
are not cached.

NB:

1. This change can only be applied to the Web-UI logic, the user still has to specify
the 'max_results' if he/she wants a lot of results.

2. /!\ The user can manually specify an arbritrary high value of 'start', so Gakusci
can ask external web-services for a huge volume of data at once, unless there already cached. /!\