External client tests are set but not activated [test]
ArxivClientTest, HalClientTest and OpenLibraryClientTest are defined
in order to check if the interaction between the client and the external
service works properly. However, those tests may fail if the related external
service is down. Because of that, they can be considered flaky.

That is why their @Test annotation are commented so they cannot be run
with Gradle (`gradle test`).

You still can activate and run those tests but only for development purpose.