Implement the Penguin Random House search client
The search service provide a simple REST API that can be used by the client.
However, the response format (in JSON or XML) cannot be deserialized properly in order to
get ISBNs.

> JSON document

Some keys (relatedisbns, title) have a value that is not consistent (this value can either an object or an array)
in the entire structure. Consequently, the deserialization of the JSON document fails.

> XML document

Jackson cannot deserialize lists to a collection properly, due to some limitations.
See https://github.com/FasterXML/jackson-dataformat-xml#known-limitations

That is why the client extracts the relevant elements (ISBN) by going through the XML data retrieved
from the search service.