Reduce the load of requests among services
When a user request n entries, the reqeust is done per service.
That mean each service must return at most n entries.
This request can be very heavy for a specific service.

That is why the application balances the maximal number of entries requested by services,
depending on the number of services used.

Example:

- 4 services are used to achieve a request.
- the user wants 100 entries.

-> each service will returns at most 25 entries.

25 = 100 / 4

NB: If the number os services is greater than the number of requested entries,
then there is no "load balancing".