Define a new way of retrieving results
> Before

The old client retrieved ISBNs and send it back to the search service.
The search service removed duplications and launch request to the Book client
in order to extract information.
This implementation had a serious flaw. Indeed, it can extract the same document from
different ISBN even because some of them are related. The consequence of that is the display
of several "identical" books with several ISBNs.
This can happen when a book has several versions (paper, ebook). Each version has its own ISBN.

> Now

The client retrieves an entry (a pair of (URI/related ISBN)) so the search service can
isolate relevant URIs describing a book. By relevant URI it means any URI that has
only one related ISBN. If several URIs that share the same set
of related ISBN, only the first one is selected.

NB: For the sake of the aggregation, and in order to avoid launching to many
parallel requests to the external web-service, the random house search service
will returns only 10 entries at most.