Set new search category in web pages: manga
I set Manga search as a specific domain in order

Why?

Even if a mang is technically a book, the users who read novels and essays are different
from the users who read mangas.

NB: Some mangas can be found in the book search, but this is due to the
external services that consider them as book (and that's true).