Improve the IResultEntry interface
1. cover()

I add the cover() abstract method in the interface because this is a common
property for all documents (except research papers/articles that have no cover,
it is does not matter).

2. Fix LSP violation

In the template (culture.html), the use of thumbnailUrl as a property of the entry
violates the Liskov Substitution Principle (LSP). Indeed, the entries sent to thymeleaf
in order to be processed are generic result entries (IResultEntry), so the application
is not supposed to assume the underlying type of the entry.