Ignore trailing slashes
Ktor no longer handles routes with or without a trailing slash as the same route.

In Gakusci, no trailing slash is used in the API definition.
However, some tests set this trailing slash when the api is used for testing.
The consequence is that those tests fail.
Resolving trailing slash as a diferent URL is none of a concern for now.
Tha is why they are ignored thanks to this new feature from Ktor 1.6.0: `IgnoreTrailingSlash`.

More info:

- https://blog.jetbrains.com/ktor/2021/05/28/ktor-1-6-0-released/
- https://youtrack.jetbrains.com/issue/KTOR-372