Change version of Syndication and logback
- Syndication 2.1.1 -> 1.1.0

I reverted Syndication because the last version is compiled in at least Java 11
and the application is still built with Java 8.

- logback 1.4.5 -> 1.3.5

The 1.4.x version of logback are built with Java 11.
Because Gakusci is still built with Java 8 I need versions of logback built the same version.
Only 1.3.x version of this dependency are built with Java 8.