Penguin Random House is not used as a service for book metasearch
The service requires authentication, which is not compatible with the use of the application.

Notes:

Even though the service is not used anymore, I am considering the possibility to integrate some APIs that require authentication.
This does not mean that it will be done, but if so, I have an idea of how this could be integrated,
without exposing API keys to the world.