Set Bucket4J to 8.4.0
This commit is more or less a revert. But also an upgrade of the dependency.

Upgrading to 8.5.0+ caused a deprecation of some functions in the library.
For now the latest version is not considered stable because some deprecated function are not properly replaced.
No upgrade above 8.4.0 will be done until this issue is done.