Update Kotlin to 2.2.21
- This update changes default rules for annotation in data classes

Consequently, in order to keep the compatibility with the behaviour in previous versions,
a new compiler option was added. Thanks to this, there is no need to modify the existing code.

Useful link: See https://youtrack.jetbrains.com/issue/KT-73255

- No compilation with Kotlin v2.3.0

Kotlin v2.3.0 was released 5 days before the date of this commit.
The source code of Gakusci, in this current state, does not compile with this version.
This is due to the atomcifu dependency that failed to be compiled with Kotlin v2.3.0.
The compiling error mentioned that atomicfu can work with Kotlin 2.2.x